Avoid C-style cast in grid

+++ source/grid/grid_generator.cc
@@ -1176,22 +1176,21 @@ namespace GridGenerator
       {
         case 1:
           for (unsigned int x = 0; x <= repetitions[0]; ++x)
-            points.push_back(p1 + (double)x * delta[0]);
+            points.push_back(p1 + x * delta[0]);
           break;
 
         case 2:
           for (unsigned int y = 0; y <= repetitions[1]; ++y)
             for (unsigned int x = 0; x <= repetitions[0]; ++x)
-              points.push_back(p1 + (double)x * delta[0] +
-                               (double)y * delta[1]);
+              points.push_back(p1 + x * delta[0] + y * delta[1]);
           break;
 
         case 3:
           for (unsigned int z = 0; z <= repetitions[2]; ++z)
             for (unsigned int y = 0; y <= repetitions[1]; ++y)
               for (unsigned int x = 0; x <= repetitions[0]; ++x)
-                points.push_back(p1 + (double)x * delta[0] +
-                                 (double)y * delta[1] + (double)z * delta[2]);
+                points.push_back(p1 + x * delta[0] + y * delta[1] +
+                                 z * delta[2]);
           break;
 
         default:
@@ -1806,22 +1805,21 @@ namespace GridGenerator
       {
         case 1:
           for (unsigned int x = 0; x <= repetitions[0]; ++x)
-            points.push_back(p1 + (double)x * delta[0]);
+            points.push_back(p1 + x * delta[0]);
           break;
 
         case 2:
           for (unsigned int y = 0; y <= repetitions[1]; ++y)
             for (unsigned int x = 0; x <= repetitions[0]; ++x)
-              points.push_back(p1 + (double)x * delta[0] +
-                               (double)y * delta[1]);
+              points.push_back(p1 + x * delta[0] + y * delta[1]);
           break;
 
         case 3:
           for (unsigned int z = 0; z <= repetitions[2]; ++z)
             for (unsigned int y = 0; y <= repetitions[1]; ++y)
               for (unsigned int x = 0; x <= repetitions[0]; ++x)
-                points.push_back(p1 + (double)x * delta[0] +
-                                 (double)y * delta[1] + (double)z * delta[2]);
+                points.push_back(p1 + x * delta[0] + y * delta[1] +
+                                 z * delta[2]);
           break;
 
         default:
@@ -5040,7 +5038,7 @@ namespace GridGenerator
                       treated_vertices[vv] = true;
                       for (unsigned int i = 0; i <= Nz; ++i)
                         {
-                          double d = ((double)i) * L / ((double)Nz);
+                          double d = i * L / Nz;
                           switch (vv - i * 16)
                             {
                               case 1:
